+++ src/render_gl2.c
@@ -955,7 +955,7 @@ static void gl2_render_layer(struct graphics_data *graphics, struct video_layer 
 }
 
 static void gl2_render_cursor(struct graphics_data *graphics,
- Uint32 x, Uint32 y, Uint8 color, Uint8 lines, Uint8 offset)
+ Uint32 x, Uint32 y, Uint16 color, Uint8 lines, Uint8 offset)
 {
   struct gl2_render_data *render_data = graphics->render_data;
   GLubyte *pal_base = &render_data->palette[color * 3];
