+++ src/render_yuv1.c
@@ -66,7 +66,7 @@ static void yuv1_render_layer(struct graphics_data *graphics, struct video_layer
 }
 
 static void yuv1_render_cursor(struct graphics_data *graphics,
- Uint32 x, Uint32 y, Uint8 color, Uint8 lines, Uint8 offset)
+ Uint32 x, Uint32 y, Uint16 color, Uint8 lines, Uint8 offset)
 {
   struct yuv_render_data *render_data = graphics->render_data;
   Uint32 *pixels;
